project/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
Django==4.1.7
django-filter==23.4
easy-thumbnails==2.8.5
fonttools==4.46.0
html5lib==1.1
Pillow==9.4.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.8.0
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.3
